Move error type into separate mod
- Resolve clippy warnings
- Improve Error struct parameter names
