+++ src/lib.rs
@@ -8,67 +8,12 @@
 #[macro_use]
 extern crate lazy_static;
 extern crate regex;
+pub mod error;
 
 use std::collections::BTreeMap;
 use std::collections::BTreeSet;
-use std::fmt;
 use regex::Regex;
-
-#[derive(Debug)]
-pub enum Error {
-    RecursiveInclude {
-        file: Option<String>,
-        recursive_file: String,
-        line_num: usize,
-        include_stack: Vec<String>,
-    },
-    FileNotFound {
-        file: Option<String>,
-        not_found: String,
-        line_num: usize,
-    },
-}
-
-impl std::error::Error for Error {
-    fn description(&self) -> &str {
-        match self {
-            &Error::RecursiveInclude { .. } => "Detected recursive #include",
-            &Error::FileNotFound { .. } => "Could not find #include file",
-        }
-    }
-
-    fn cause(&self) -> Option<&std::error::Error> {
-        None
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            &Error::RecursiveInclude {
-                ref file,
-                ref recursive_file,
-                ref line_num,
-                ref include_stack,
-                ..
-            } => write!(
-                f,
-                "Detected recursive include of file \"{}\" in file {:?}, line {}, include stack {:?}",
-                recursive_file, file, line_num, include_stack
-            ),
-            &Error::FileNotFound {
-                ref not_found,
-                ref file,
-                ref line_num,
-                ..
-            } => write!(
-                f,
-                "Could not find file \"{}\", included from file {:?}, line {}\nhelp: Call Preprocessor::file with the file name and contents",
-                not_found, file, line_num
-            ),
-        }
-    }
-}
+use error::Error;
 
 #[derive(Debug, Default)]
 pub struct Preprocessor<'a> {
@@ -158,12 +103,14 @@ impl<'a> Preprocessor<'a> {
 
                 // if this file is already in our include stack, return Err
                 if include_stack.contains(&file) {
-                    let stack = include_stack.into_iter().map(|s| s.to_string()).collect();
+                    let in_file = name.map(|s| s.to_string());
+                    let problem_include = file.to_string();
+                    let include_stack = include_stack.into_iter().map(|s| s.to_string()).collect();
                     return Err(Error::RecursiveInclude {
-                        file: name.map(|s| s.to_string()),
-                        recursive_file: file.to_string(),
+                        in_file: in_file,
                         line_num: line_num,
-                        include_stack: stack,
+                        problem_include: problem_include,
+                        include_stack: include_stack,
                     });
                 }
                 include_stack.push(&file);
@@ -179,10 +126,12 @@ impl<'a> Preprocessor<'a> {
                         include_set,
                     )?;
                 } else {
+                    let in_file = name.map(|s| s.to_string());
+                    let problem_include = file.to_string();
                     return Err(Error::FileNotFound {
-                        file: name.map(|s| s.to_string()),
-                        not_found: file.to_string(),
+                        in_file: in_file,
                         line_num: line_num,
+                        problem_include: problem_include,
                     });
                 }
                 include_stack.pop();
